BLE Scanning: Force the APCF address type value to always be 0x02.
The address type in...

Java:
 Public = 0
 Random = 1

In the host stack when dealing with BLE ID addresses 2 and 3 are used for resolve
address
C++:
 Public = 0
 Random = 1
 Public ID = 2
 Random ID = 3

The vendor command takes Public (0) or Random (1) or Any (2)

We need to use 0x02 for Any if we want to use the resolving list for scans.

e.g. specifying for a PUBLIC (0x00) will only return results for a public matching
address and will NOT try to resolve any RPA to the address.  For this the Any (0x02)
must be specified.

Ignore-AOSP-First: Need on sc-v2-dev; code doesn't match on master

Bug: 209885517
Test: Manual scanning app
Tag: #stability
Change-Id: Ic4236d5703611de83c45ebe7638c0f4f2b02931b